requirements.txt
attrs==21.2.0
bitmex==0.2.2
bravado==11.0.3
bravado-core==5.17.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
monotonic==1.6
msgpack==1.0.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3987==1.3.8
simplejson==3.17.2
six==1.16.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.4
webcolors==1.11.1
Werkzeug==2.0.1
